Disk IO library: fix integer overflow in IO buffer pool get method with buffer pool larger than 4GB

+++ src/cc/qcdio/QCIoBufferPool.cc
@@ -123,7 +123,7 @@ public:
         QCASSERT(theIdx > 0);
         mFreeCnt--;
         *mFreeListPtr = mFreeListPtr[theIdx];
-        return (mStartPtr + ((theIdx - 1) << mBufSizeShift));
+        return (mStartPtr + (size_t(theIdx - 1) << mBufSizeShift));
     }
 
     bool Put(char* inPtr)
